requirements.txt
openai>=1.68.2,<1.76.0
python-dotenv==1.0.0
flask==3.0.0 